Add configuration, map, view, widgets, layers, and utility modules
* **src/config.js**: Export `esriConfig` object with API key.
* **src/map.js**: Import necessary modules and export `map` object with basemap and layers.
* **src/view.js**: Import necessary modules and export `activeView` object with map and view configuration.
* **src/widgets.js**: Import necessary modules and export `widgets` object with widget configurations.
* **src/layers.js**: Import necessary modules and export `layers` object with layer configurations.
* **src/utils.js**: Export utility functions for handling layer view clicks and updating the PDF iframe.
* **src/main.js**: Refactor to import and use the new modules and utility functions.
